+++ packages/react-native-quick-crypto/src/pbkdf2.ts
@@ -1,15 +1,19 @@
-import { NativeQuickCrypto } from './NativeQuickCrypto/NativeQuickCrypto';
 import { Buffer } from '@craftzdog/react-native-buffer';
-import {
-  type BinaryLike,
-  binaryLikeToArrayBuffer,
-  lazyDOMException,
-  bufferLikeToArrayBuffer,
-  normalizeHashName,
-  HashContext,
+import { NitroModules } from 'react-native-nitro-modules';
+import type {
+  BinaryLike,
 } from './utils';
-import type { CryptoKey, HashAlgorithm, SubtleAlgorithm } from './keys';
+import {
+  HashContext,
+  binaryLikeToArrayBuffer,
+  bufferLikeToArrayBuffer,
+  lazyDOMException,
+  normalizeHashName
+} from './utils';
+import type { HashAlgorithm, SubtleAlgorithm } from './utils';
+import type { Pbkdf2 } from './specs/pbkdf2.nitro';
 import { promisify } from 'util';
+import type { CryptoKey } from './keys';
 
 const WRONG_PASS =
   'Password must be a string, a Buffer, a typed array or a DataView';
@@ -19,66 +23,47 @@ type Password = BinaryLike;
 type Salt = BinaryLike;
 type Pbkdf2Callback = (err: Error | null, derivedKey?: Buffer) => void;
 
+// to use native bits in sub-functions, use getNative(). don't call it at top-level!
+let native: Pbkdf2;
+function getNative(): Pbkdf2 {
+  if (native == null) {
+    // lazy-load the Nitro HybridObject
+    native = NitroModules.createHybridObject<Pbkdf2>('Pbkdf2');
+  }
+  return native;
+}
+
 function sanitizeInput(input: BinaryLike, errorMsg: string): ArrayBuffer {
   try {
     return binaryLikeToArrayBuffer(input);
-  } catch (e: any) {
-    throw errorMsg;
+    // eslint-disable-next-line @typescript-eslint/no-unused-vars
+  } catch (_e: unknown) {
+    throw new Error(errorMsg);
   }
 }
 
-const nativePbkdf2 = NativeQuickCrypto.pbkdf2;
-
 export function pbkdf2(
   password: Password,
   salt: Salt,
   iterations: number,
   keylen: number,
-  digest: HashAlgorithm,
-  callback: Pbkdf2Callback
-): void;
-export function pbkdf2(
-  password: Password,
-  salt: Salt,
-  iterations: number,
-  keylen: number,
-  callback: Pbkdf2Callback
-): void;
-export function pbkdf2(
-  password: Password,
-  salt: Salt,
-  iterations: number,
-  keylen: number,
-  arg0?: unknown,
-  arg1?: unknown
+  digest: string,
+  callback: Pbkdf2Callback,
 ): void {
-  let digest: HashAlgorithm = 'SHA-1';
-  let callback: undefined | Pbkdf2Callback;
-  if (typeof arg0 === 'string') {
-    digest = arg0 as HashAlgorithm;
-    if (typeof arg1 === 'function') {
-      callback = arg1 as Pbkdf2Callback;
-    }
-  } else {
-    if (typeof arg0 === 'function') {
-      callback = arg0 as Pbkdf2Callback;
-    }
-  }
-  if (callback === undefined) {
+  if (callback === undefined || typeof callback !== 'function') {
     throw new Error('No callback provided to pbkdf2');
   }
-
   const sanitizedPassword = sanitizeInput(password, WRONG_PASS);
   const sanitizedSalt = sanitizeInput(salt, WRONG_SALT);
   const normalizedDigest = normalizeHashName(digest, HashContext.Node);
 
-  nativePbkdf2
-    .pbkdf2(
+  getNative();
+  native.pbkdf2(
       sanitizedPassword,
       sanitizedSalt,
       iterations,
       keylen,
-      normalizedDigest
+      normalizedDigest,
     )
     .then(
       (res: ArrayBuffer) => {
@@ -86,7 +71,7 @@ export function pbkdf2(
       },
       (e: Error) => {
         callback!(e);
-      }
+      },
     );
 }
 
@@ -101,7 +86,8 @@ export function pbkdf2Sync(
   const sanitizedSalt = sanitizeInput(salt, WRONG_SALT);
 
   const algo = digest ? normalizeHashName(digest, HashContext.Node) : 'sha1';
-  let result: ArrayBuffer = nativePbkdf2.pbkdf2Sync(
+  getNative();
+  const result: ArrayBuffer = native.pbkdf2Sync(
     sanitizedPassword,
     sanitizedSalt,
     iterations,
@@ -152,7 +138,7 @@ export async function pbkdf2DeriveBits(
 
   const sanitizedPassword = sanitizeInput(raw, WRONG_PASS);
   const sanitizedSalt = sanitizeInput(salt, WRONG_SALT);
-  let result: Buffer | undefined = await pbkdf2Promise(
+  const result: Buffer | undefined = await pbkdf2Promise(
     sanitizedPassword,
     sanitizedSalt,
     iterations,
