+++ packages/react-native-quick-crypto/src/utils/errors.ts
@@ -1,0 +1,15 @@
+type DOMName =
+  | string
+  | {
+      name: string;
+      cause: unknown;
+    };
+
+export function lazyDOMException(message: string, domName: DOMName): Error {
+  let cause = '';
+  if (typeof domName !== 'string') {
+    cause = `\nCaused by: ${domName.cause}`;
+  }
+
+  return new Error(`[${domName}]: ${message}${cause}`);
+}
