+++ packages/react-native-quick-crypto/src/utils/index.ts
@@ -1,6 +1,5 @@
-import { ab2str, abvToArrayBuffer } from "./conversion";
+export * from './conversion';
+export * from './errors';
+export * from './hashnames';
+export * from './types';
 
-export const utils = {
-  ab2str,
-  abvToArrayBuffer,
-};
