+++ packages/react-native-quick-crypto/zzz/ts/NativeQuickCrypto/webcrypto.ts
@@ -16,15 +16,6 @@ import type {
 import type { KeyVariant } from './Cipher';
 import type { RSACipher } from './rsa';
 
-type KeyDetail = {
-  length?: number;
-  publicExponent?: number;
-  modulusLength?: number;
-  hashAlgorithm?: string;
-  mgf1HashAlgorithm?: string;
-  saltLength?: number;
-  namedCurve?: string;
-};
 
 type ECExportKey = (
   format: KWebCryptoKeyFormat,
@@ -37,26 +28,7 @@ type RSAExportKey = (
   variant: KeyVariant
 ) => ArrayBuffer;
 
-export type KeyObjectHandle = {
-  export(
-    format?: KFormatType,
-    type?: KeyEncoding,
-    cipher?: string,
-    passphrase?: ArrayBuffer
-  ): ArrayBuffer;
-  exportJwk(key: JWK, handleRsaPss: boolean): JWK;
-  getAsymmetricKeyType(): AsymmetricKeyType;
-  init(
-    keyType: KeyType,
-    key: any,
-    format?: KFormatType,
-    type?: KeyEncoding,
-    passphrase?: string | ArrayBuffer
-  ): boolean;
-  initECRaw(curveName: string, keyData: ArrayBuffer): boolean;
-  initJwk(keyData: JWK, namedCurve?: NamedCurve): KeyType | undefined;
-  keyDetail(): KeyDetail;
-};
+
 
 type CreateKeyObjectHandle = () => KeyObjectHandle;
